MPI ABI: Fix Comm and Errhandler behaviour
The recent change in MPI_Comm (mostly the constants) broke the
errhandler retrievement. It has been fixed using the mapping functions
created to deal with these constants.

Furthermore, communicator could be created without errhandler and no
fallback existed. Now, communicators inherit of the errhandler of their
parent on creation as a default behaviour to properly call errhandlers.

Comments has also been added.

+++ src/MPC_MPI/src/errh.h
@@ -50,7 +50,7 @@ _mpc_mpi_errhandler_t _mpc_mpi_errhandler_from_idx (const long idx);
  *  @param errh  The errhandler of which we want the index in the global table
  *  @return      A valid index
  */
-unsigned short _mpc_mpi_errhandler_to_idx (const _mpc_mpi_errhandler_t errh);
+long _mpc_mpi_errhandler_to_idx (const _mpc_mpi_errhandler_t errh);
 
 /** @brief Checks if an errhandler is valid (registered)
  *
